cleaning code on allocation plan persistence
