+++ src/main/go/domain/infra/repository/allocation_plan_repository_model.go
@@ -1,0 +1,78 @@
+package repository
+
+import (
+	"github.com/benizzio/open-asset-allocator/domain"
+	"github.com/benizzio/open-asset-allocator/domain/allocation"
+	"github.com/benizzio/open-asset-allocator/infra/sqlext"
+	"github.com/shopspring/decimal"
+)
+
+type plannedAllocationJoinedRowDTS struct {
+	AllocationPlanId     int
+	Name                 string
+	Type                 allocation.PlanType
+	PlannedExecutionDate sqlext.NullTime
+	StructuralId         sqlext.NullStringSlice
+	CashReserve          bool
+	SliceSizePercentage  decimal.Decimal
+}
+
+func mapPlannedAllocationRows(rows []plannedAllocationJoinedRowDTS) ([]*domain.AllocationPlan, error) {
+
+	var allocationPlanCacheMap = make(map[int]*domain.AllocationPlan)
+	var allocationPlans = make([]*domain.AllocationPlan, 0)
+
+	for _, row := range rows {
+
+		allocationPlan := mapPlannedAllocationRow(&row, allocationPlanCacheMap)
+		if allocationPlan != nil {
+			allocationPlans = append(allocationPlans, allocationPlan)
+		}
+	}
+
+	return allocationPlans, nil
+}
+
+func mapPlannedAllocationRow(
+	row *plannedAllocationJoinedRowDTS,
+	allocationPlanCacheMap map[int]*domain.AllocationPlan,
+) *domain.AllocationPlan {
+
+	allocationPlanUnit := buildPlannedAllocationFromRow(row)
+
+	if cachedAllocationPlan, exists := allocationPlanCacheMap[row.AllocationPlanId]; exists {
+		cachedAllocationPlan.AddDetail(allocationPlanUnit)
+	} else {
+		allocationPlan := buildAllocationPlanFromRow(row, allocationPlanUnit)
+		allocationPlanCacheMap[row.AllocationPlanId] = allocationPlan
+		return allocationPlan
+	}
+
+	return nil
+}
+
+func buildPlannedAllocationFromRow(rowDTS *plannedAllocationJoinedRowDTS) *domain.PlannedAllocation {
+	return &domain.PlannedAllocation{
+		StructuralId:        rowDTS.StructuralId.ToStringSlice(),
+		CashReserve:         rowDTS.CashReserve,
+		SliceSizePercentage: rowDTS.SliceSizePercentage,
+	}
+}
+
+func buildAllocationPlanFromRow(
+	rowDTS *plannedAllocationJoinedRowDTS,
+	plannedAllocation *domain.PlannedAllocation,
+) *domain.AllocationPlan {
+
+	var allocationPlan = domain.AllocationPlan{
+		AllocationPlanIdentifier: domain.AllocationPlanIdentifier{
+			Id:   rowDTS.AllocationPlanId,
+			Name: rowDTS.Name,
+		},
+		PlanType:             rowDTS.Type,
+		PlannedExecutionDate: rowDTS.PlannedExecutionDate.ToTimeReference(),
+	}
+	allocationPlan.AddDetail(plannedAllocation)
+
+	return &allocationPlan
+}
